added several new components
