+++ client/src/ReactRouter.js
@@ -1,9 +1,9 @@
 import React from "react";
-import Home from "./pages/Home/Home";
-import Store from "./pages/Store/Store";
-import Product from "./pages/Product/Product";
-import Cart from "./pages/Cart/Cart";
-import NoMatch from "./pages/NoMatch/NoMatch";
+import Home from "./pages/Home";
+import Store from "./pages/Store";
+import ProductDetails from "./pages/ProductDetails";
+import Cart from "./pages/Cart";
+import NoMatch from "./pages/NoMatch";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 const ReactRouter = () => (
@@ -13,7 +13,7 @@ const ReactRouter = () => (
 			<Route exact path="/home" component={Home} />
 			<Route exact path="/store" component={Store} />
 			<Route exact path="/cart" component={Cart} />
-			<Route exact path="/product/:productId" render={props => (<Product routeProps={props} />)} />
+			<Route exact path="/product/:productId" render={props => (<ProductDetails routeProps={props} />)} />
 
 			<Route component={NoMatch} />
 		</Switch>
