--- client/src/pages/Store/store.css
@@ -1,24 +1,0 @@
-.catItem {
-  cursor: pointer;
-  background-color: transparent;
-  color: grey;
-  border-style: none;
-  padding: 10px;
-  text-align: center;
-  text-decoration: none;
-  border-radius: 5px;
-  border-bottom: 5px solid transparent;
-}
-
-.catItem:hover {
-  background-color: rgb(190, 245, 190);
-  color: darkgreen;
-}
-
-.activeItem {
-  border-radius: 5px;
-  background-color: rgb(190, 245, 190);
-  color: darkgreen;
-  border-bottom: 5px solid darkgreen;
-  font-weight: bold;
-}
